Only alias the EntityManager class if it doesn't yet exist.
See #17, when running unit tests, this file is included multiple times. The second time, the `EntityManager` class will already exist, so `class_alias` will fail with the following error:
```
Cannot declare class EntityManager, because the name is already in use
```

Wrapping the statement in a `class_exists` condition fixes this.